extracted wavewrite from convenience, improved rtl_fm
* rtl_fm using rtlsdr_set_tuner_band_center
* rtl_fm recording to .wav.tmp, renaming it when finished
